requirements.txt
beautifulsoup4
trafilatura
langchain
langchain-community
langchain-openai
tqdm
faiss-cpu
openai
python-dotenv
tiktoken
